requirements.txt
beautifulsoup4==4.10.0
discord.py==1.7.3
requests==2.25.1
requests-oauthlib==1.3.0
async-generator==1.10
async-timeout==3.0.1
aiohttp==3.7.4.post0
Pillow==8.2.0
gpt-2-simple==0.7.2
gdown==3.14.0
nltk==3.6.2
geograpy3==0.2.1
tensorflow==1.15.4
pandas==1.4.1
dataframe-image==0.1.1
